requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
bitarray==2.9.2
cachetools==4.2.4
certifi==2024.2.2arn
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cytoolz==0.12.3
Django==4.2.11
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.12.0
hexbytes==0.3.1
httpcore==0.14.7
idna==3.6
importlib_metadata==7.0.2
inflection==0.5.1
ipfshttpclient==0.8.0a2
jsonalias==0.1.1
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
lru-dict==1.3.0
Markdown==3.5.2
multiaddr==0.0.9
multidict==6.0.5
netaddr==1.2.1
packaging==23.2
parsimonious==0.8.1
pillow==10.2.0
protobuf==3.20.3
psycopg2==2.9.9
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
PyNaCl==1.5.0
pypng==0.20220715.0
pyrsistent==0.20.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.33.0
requests==2.31.0
rfc3986==1.5.0
rlp==2.0.1
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
solders==0.20.0
sqlparse==0.4.4
toolz==0.12.1
types-cachetools==4.2.10
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
varint==1.0.2
web3==5.28.0
websockets==9.1
yarl==1.9.4
zipp==3.17.0
